pip-requirements.txt
Pillow==6.0.0
pkg-resources==0.0.0
pyusb==1.0.2
PyVISA==1.9.1
PyVISA-py==0.3.1
